+++ hyperspy/_signals/image.py
@@ -148,3 +148,20 @@ class Image(Signal):
             centre_colormap=centre_colormap,
             **kwargs
         )
+
+    def create_model(self, dictionary=None):
+        """Create a model for the current signal
+
+        Parameters
+        __________
+        dictionary : {None, dict}, optional
+            A dictionary to be used to recreate a model. Usually generated using
+            :meth:`hyperspy.model.as_dictionary`
+
+        Returns
+        -------
+        A Model class
+
+        """
+        from hyperspy.model import Model2D
+        return Model2D(self, dictionary=dictionary)
